poc/apps/mcp_server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "MCP Server application"
requires-python = ">=3.14"
dependencies = [
    "duckdb>=1.4.1",
    "duckdb-engine>=0.17.0",
    "mcp>=1.0.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.43",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server"]
